factors: Add more tests to factor-outbound-pg and refactor it to be generic across pg impl
